refactor(cli): extract config and permission commands
